+++ src/lib/openapi/spec/custom-metrics-schema.ts
@@ -1,36 +1,6 @@
 import type { FromSchema } from 'json-schema-to-ts';
 import { dateSchema } from './date-schema.js';
-
-export const customMetricSchema = {
-    $id: '#/components/schemas/customMetricSchema',
-    type: 'object' as const,
-    required: ['name', 'value'],
-    description: 'A custom metric with name, value and optional labels',
-    properties: {
-        name: {
-            type: 'string' as const,
-            description: 'Name of the custom metric',
-            example: 'http_responses_total',
-        },
-        value: {
-            type: 'number' as const,
-            description: 'Value of the custom metric',
-            example: 1,
-        },
-        labels: {
-            type: 'object' as const,
-            description: 'Labels to categorize the metric',
-            additionalProperties: {
-                type: 'string' as const,
-            },
-            example: {
-                status: '200',
-                method: 'GET',
-            },
-        },
-    },
-    components: {},
-};
+import { customMetricSchema } from './custom-metric-schema.js';
 
 export const customMetricsSchema = {
     $id: '#/components/schemas/customMetricsSchema',
@@ -54,5 +24,4 @@ export const customMetricsSchema = {
     },
 } as const;
 
-export type CustomMetricSchema = FromSchema<typeof customMetricSchema>;
 export type CustomMetricsSchema = FromSchema<typeof customMetricsSchema>;
